+++ src/main/java/drzhark/mocreatures/client/model/MoCModelOstrich.java
@@ -1143,8 +1143,8 @@ public class MoCModelOstrich extends ModelBase {
     ModelRenderer Tailpart4;
     ModelRenderer Tailpart5;
 
-    private byte helmet;
+    private int helmet;
     private byte typeI;
-    private byte flagColor;
+    private int flagColor;
     private float radianF = 57.29578F;
 }
